Add new route for measure and measurement
