+++ app/serializers/measure_serializer.rb
@@ -1,0 +1,6 @@
+class MeasureSerializer
+  include FastJsonapi::ObjectSerializer
+  attributes :title, :description, :image
+
+  has_many :measurements
+end
