+++ app/serializers/measurement_serializer.rb
@@ -1,0 +1,4 @@
+class MeasurementSerializer
+  include FastJsonapi::ObjectSerializer
+  attributes :value, :date, :measure_id, :user_id
+end
